Add files via upload
This project focused on assigning probabilities to customers in telecommunication sector using Logistic Regression Model in order to determine their propensity to churn. If the predicted probability was 0.5 or better, the algorithm classified the customer as belonging to Churn class and if the assigned probability was less than 0.5, then the customer was classified as a non-churner. A Theoretical framework called Push, Pull and Mooring (PPM) Framework was employed to explain churn behaviour among customers with highest propensities to churn and it was established that Push Factors (dissatisfaction effects with the current service provider) were dominant among high risk customers. Significant Factors explaining churn behaviour in the Reduced Logistic Regression Model were also determined.